backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.2

# Redis
redis[hiredis]==5.1.0

# AI / LLM
anthropic==0.39.0
openai==1.50.0
langchain==0.3.0
langgraph==0.2.0
tiktoken==0.7.0

# Vector DB & Embeddings
pymilvus==2.4.5

# Document Processing
python-docx==1.1.2
python-pptx==1.0.2
openpyxl==3.1.5
pypdf==4.3.1
pdfplumber==0.11.4

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.9.0
pydantic-settings==2.5.0

# Object Storage
boto3==1.35.0

# Utilities
httpx==0.27.2
tenacity==9.0.0
structlog==24.4.0
numpy==1.26.4
